analyser/requirements.txt
fasttext==0.9.2
Flask==2.0.1
nltk==3.6.2
numpy==1.19.5
#psycopg2==2.7.7
psycopg2-binary
python-dotenv==0.19.0